requirements.txt
# PDF Processing
pdfplumber>=0.10.0
PyPDF2>=3.0.0
tabula-py>=2.8.0

# NLP
spacy>=3.7.0
nltk>=3.8.0
scikit-learn>=1.3.0
# gensim>=4.3.0  # Temporarily disabled - compilation issues with Python 3.13
# Note: LDA functionality will use scikit-learn's LatentDirichletAllocation instead

# Data Processing
pandas>=2.1.0
numpy>=1.25.0
scipy>=1.11.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0
wordcloud>=1.9.3

# Streamlit & UI
streamlit>=1.30.0
streamlit-aggrid>=0.3.4
streamlit-option-menu>=0.3.6
# streamlit-extras>=0.3.6  # Temporarily disabled for Python 3.13 compatibility

# Utils
pyyaml>=6.0
python-dotenv>=1.0.0
Pillow>=10.0.0
openpyxl>=3.1.2
tqdm>=4.66.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.1.0
reportlab>=4.0.0  # For PDF generation in tests
